fix: better code
